requirements.txt
# -----------------------------------------------------------------------------
# Financial Insight Tools (FIT) — runtime dependencies
#
# FIT is a Streamlit-based, modular research environment.
# Requirements are intentionally minimal and pinned to stable ranges.
# -----------------------------------------------------------------------------

# Core UI framework
streamlit>=1.51,<1.60

# Data layer
pandas>=2.2,<2.4
numpy>=2.0,<2.2
scipy>=1.15,<1.16

# Visualisation
plotly>=5.22,<6.0
altair>=5.5,<6.0

# YAML configuration / references
PyYAML>=6.0,<7.0

# Maps / geo
folium>=0.19,<0.21
streamlit-folium>=0.24,<0.25

# Tables / grids
streamlit-aggrid>=1.1,<1.2
